ai_service/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.3
groq==0.4.1
Pillow==10.4.0
psycopg2-binary==2.9.9
